.history/requirements_20251004124605.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web Framework & API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.28.0
requests>=2.31.0

# Vector Databases & Embeddings
chromadb>=0.4.15
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
langchain>=0.0.350
openai>=1.3.0

# Database Connectors
mysql-connector-python>=8.1.0
psycopg2-binary>=2.9.0

# Text Processing
nltk>=3.8.0
regex>=2023.0.0

# System Utilities
psutil>=5.9.0
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.4.0
black>=23.0.0